Move styled components for shared components to styled files
