+++ src/components/Header/index.tsx
@@ -1,28 +1,12 @@
 import * as React from 'react';
-import { Link } from 'gatsby';
-import styled, { ThemeProvider, useTheme } from 'styled-components';
-import { AppBar, Box, Toolbar, Typography } from '@mui/material';
+import { ThemeProvider, useTheme } from 'styled-components';
+import { Box, Toolbar, Typography } from '@mui/material';
 
-import { appBarTheme } from '../themes';
-
-const StyledAppBar = styled(AppBar)`
-    background-color: ${(props) => props?.theme?.backgroundColor} !important;
-    color: ${(props) => props.theme?.color} !important;
-    z-index: 1000000 !important;
-`;
-
-const NavLink = styled(Link)`
-    color: ${(props) => props?.theme?.color};
-    font-size: 1.5rem;
-    margin-left: 12em;
-    padding: 0 1em;
-    text-align: center;
-    text-decoration: none;
-`;
+import { appBarTheme } from '../../themes';
+import { StyledAppBar, NavLink } from './styled';
 
 const Header = ({ data }: HeaderProps): React.ReactElement => {
     const theme = useTheme();
-
 
     // console.log('---------- data: ', data);
     // console.log('---------- nodes: ', nodes);
