+++ src/components/Header/styled.tsx
@@ -1,0 +1,20 @@
+import styled from 'styled-components';
+import { Link } from 'gatsby';
+import { AppBar } from '@mui/material';
+
+const StyledAppBar = styled(AppBar)`
+    background-color: ${(props) => props?.theme?.backgroundColor} !important;
+    color: ${(props) => props.theme?.color} !important;
+    z-index: 1000000 !important;
+`;
+
+const NavLink = styled(Link)`
+    color: ${(props) => props?.theme?.color};
+    font-size: 1.5rem;
+    margin-left: 12em;
+    padding: 0 1em;
+    text-align: center;
+    text-decoration: none;
+`;
+
+export { StyledAppBar, NavLink };
